splitting up some classes
